requirements.txt
attrs==21.4.0
boto3==1.24.7
botocore==1.27.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
dc-python-sdk==0.0.4
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
jmespath==1.0.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycryptodome==3.14.1
pyparsing==3.0.8
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.0
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
